+++ src/components/themeTypes.ts
@@ -11,6 +11,8 @@ import type { ListTheme } from './ui/typography/list/ListTheme';
 import type { ListItemTheme } from './ui/typography/listItem/ListItemTheme';
 import type { CardTheme } from "./ui/card/CardTheme";
 import type { FieldTheme } from "./ui/field/FieldTheme";
+import type { AlertTheme } from "./ui/alert/AlertTheme";
+import type { SpinnerTheme } from "./ui/spinner/SpinnerTheme";
 import type { CardHeaderTheme } from './ui/card/CardHeaderTheme';
 import type { CardBodyTheme } from './ui/card/CardBodyTheme';
 import type { CardFooterTheme } from './ui/card/CardFooterTheme';
@@ -66,6 +68,8 @@ import type { ButtonProps } from "./ui/button/ButtonProps";
 import type { ButtonSpinnerProps } from "./ui/button/ButtonSpinnerProps";
 import type { CardProps } from "./ui/card/CardProps";
 import type { FieldProps } from "./ui/field/FieldProps";
+import type { AlertProps } from "./ui/alert/AlertProps";
+import type { SpinnerProps } from "./ui/spinner/SpinnerProps";
 import type { CardHeaderProps } from "./ui/card/CardHeaderProps";
 import type { CardBodyProps } from "./ui/card/CardBodyProps";
 import type { CardFooterProps } from "./ui/card/CardFooterProps";
@@ -194,6 +198,9 @@ export interface ThemeProps {
     closeButton: ComponentTheme<ModalCloseButtonProps, ModalCloseButtonTheme>;
   };
   popup: ComponentTheme<PopupProps, PopupTheme>;
+  tooltip: ComponentTheme<PopupProps, PopupTheme>;
+  alert: ComponentTheme<AlertProps, AlertTheme>;
+  spinner: ComponentTheme<SpinnerProps, SpinnerTheme>;
   menu: {
     item: ComponentTheme<MenuItemProps, MenuItemTheme>;
     popup: ComponentTheme<PopupProps, PopupTheme>;